src/cluster-mgmt/src/cluster-mgmt-exporter/requirements.txt
# python==3.8
prometheus-client==0.15.0
pyyaml
pyNfsClient
kubernetes==24.2.0
asyncssh==2.14.2
aiohttp
pytimeparse2==1.7.1
# node exporter log tailing
pygtail==0.14.0
# system client
protobuf==5.27.2
types-protobuf==5.26.0.20240422
grpcio==1.64.1
grpcio-tools==1.64.1
mypy-protobuf==3.6.0
absl-py==2.1.0
paramiko==3.4.0
sshtunnel==0.4.0
urllib3<2
# for testing
pytest>=8.2.1
pytest-mock>=3.14.0
# fastapi related
fastapi==0.111.0
uvicorn==0.30.1
starlette==0.37.2
psutil==6.0.0
pytest-asyncio==0.23.7
pymongo==3.10.1
aioresponses
